==> lecture12-simple-linear-regression/00_life_exp_get_data.ipynb <==
# Lecture 12

## Getting the data for analysis
   - practice with WDI package           
                                           
#### Case Study:                               
   - Life-expectancy and income   
---

import pandas as pd
import numpy as np
from datetime import date
import warnings

warnings.filterwarnings("ignore")

import wbdata

Reminder on how WDI works - it is an API\
Search for variables which contains GDP

wbdata.search_indicators("gdp")

Narrow down the serach for: GDP + something + capita + something + constant

wbdata.search_indicators("gdp.*capita.*constant")

Get GDP data

gdp_data = wbdata.get_data(
    indicator="NY.GDP.PCAP.PP.KD", country="all", data_date=date(2019, 1, 1)
)
gdp_data

### Task: 

Get the GDP data, along with `population`, `total` and `life expectancy at birth`
for year 2019 and save to your data folder!\
Note: We have pushed it to Github, we will use that later, just to be on the same page!

wbdata.search_indicators('population, total')

wbdata.search_indicators("life expectancy at birth")

data_raw = wbdata.get_dataframe(
    indicators={
        "NY.GDP.PCAP.PP.KD": "gdppc",
        "SP.DYN.LE00.IN": "lifeexp",
        "SP.POP.TOTL": "population",
    },
    country="all",
    data_date=date(2019, 1, 1),
).reset_index()

ISOcodes = pd.DataFrame(wbdata.get_country())[["iso2Code", "name"]]

data_raw = (
    data_raw.reset_index(drop=True)
    .merge(ISOcodes, left_on="country", right_on="name", how="left")
    .drop("name", axis=1)
)

data_raw.to_csv("data/WDI_lifeexp_raw.csv", index = False)
