==> lecture12-simple-linear-regression/01_life_exp_clean.ipynb <==
#               Lecture 12                  #
                                          
##  Auxiliary file to clean data            
    - can practice, but not mandatory   
                                          
#### Case Study:                               
- life-expectancy and income               
---

import pandas as pd
import numpy as np
import warnings

warnings.filterwarnings("ignore")

Call the data from github

my_url = "https://raw.githubusercontent.com/gabors-data-analysis/da-coding-python/main/lecture12-simple-linear-regression/data/WDI_lifeexp_raw.csv"
#df = pd.read_csv(my_url)
df = pd.read_csv("data/WDI_lifeexp_raw.csv")
df.loc[lambda x: x["country"] == "Namibia", "iso2Code"] = "NA"
df = df.dropna(subset=["iso2Code"])

Check the observations:

Lot of grouping observations usually contains a number

d1 = df.loc[~df["iso2Code"].str.isalpha()]
d1

Filter these out

df = df.loc[df["iso2Code"].str.isalpha()]

 Some grouping observations are still there, check each of them\
   HK - Hong Kong, China\
   OE - OECD members\
   all with starting X, except XK which is Kosovo\
   all with starting Z, except ZA-South Africa, ZM-Zambia and ZW-Zimbabwe\

1st drop speficif values

df.loc[lambda x: x["iso2Code"].isin(["EU","HK","OE"])]

Save opposite

df = df.loc[lambda x: ~x["iso2Code"].isin(["EU","HK","OE"])]

2nd drop values with certain starting character\
Get the first letter from iso2c

df = df.loc[
    (~((df["iso2Code"].str[0] == "X") | (df["iso2Code"].str[0] == "Z")))
    | (df["iso2Code"].isin(["XK", "ZA", "ZM", "ZW"]))
]

Check for missing observations

df = df.dropna(subset=['gdppc', 'lifeexp', 'population'])

### Clean variables

Rename variables and scale them\
Drop all the others !! in this case write into readme it is referring to year 2018!!

df["population"] = df["population"] / 10**6
df["gdppc"] = df["gdppc"] / 10**3

df.filter([ "population", "gdppc", "lifeexp"]).hist()

It seems we have a large value(s) for population:

df.loc[df["population"]>500]

These are India and China... not an extreme value

df.describe()

Save the raw data file for your working directory

df.to_csv("data/WDI_lifeexp_clean.csv",index=False)
